refactor(boot): give the exit path its own file
How the program ends is a concern of its own: the exit report, the fatal-error
dialog, the two LBA_NORETURN exits and the boot-time failure path. It sat in
PERSO.CPP, which is the main loop and the entity framework, because that is where
main() has been since 1997 and boot work has accreted around it ever since.

BOOT_EXIT.CPP takes it, and PERSO.CPP goes from 3206 lines to 3007. A surface in
CODESTYLE's sense: everything calls it, it owns nobody else's state, and nothing
in it reaches for a subsystem's variables.

The exit report turns out to be private already. End_Num and End_Error were
file-scope statics in PERSO.CPP, so they move with the block and an exit is now
decided by calling the module rather than by assigning its state.

Two sites in main() assigned End_Error directly, which is the one leak the
extraction was expected to find. BootExit_NoteContext takes both: boot names the
config file on its way past, so a failure before anything more specific points at
the file being read rather than at nothing.

A move, not a change: the block is the same lines in a new home, and a start that
fails on missing data prints byte-identical output and the same exit code. The
em-dash guard is bypassed for that reason, on two comments that came across
verbatim; they are rewritten in the next commit rather than inside the move.

Deliberately left in PERSO.CPP: main() itself, the platform bring-up around it,
and the asset loading. The roadmap says why. The entry point moves when a game-id
exists to say what boot should call instead.
